% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{cdcs_dict}
\alias{cdcs_dict}
\title{CIDACS Dictionary}
\usage{
cdcs_dict
}
\description{
A data dictionary for CIDACS data, which can be found \href{https://gcgh.grandchallenges.org/sites/default/files/additional-materials/CIDACS Dicionario Dados Coorte 100M.PDF}{here}.
}
