% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{clean_data}
\alias{clean_data}
\title{Clean DATASUS data}
\usage{
clean_data(data)
}
\arguments{
\item{data}{a data frame}
}
\description{
Clean DATASUS data
}
\examples{
d <- transform_data(sinasc_example, subset(cdcs_dict, db == "SINASC"))
clean_data(d)

\dontrun{
fin <- "ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/NOV/DNRES/DNRO2013.DBC"
fout <- tempfile(fileext = ".DBC")
download.file(fin, destfile = fout)
d <- read_datasus(fout)
d <- transform_data(d, subset(cdcs_dict, db == "SINASC"))
d <- clean_data(d)
}
}
