% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_datasus}
\alias{read_datasus}
\title{Read data from a DATASUS source (SINASC/SIM)}
\usage{
read_datasus(path)
}
\arguments{
\item{path}{a path to a dbf, dbc, or csv file}
}
\description{
Read data from a DATASUS source (SINASC/SIM)
}
\examples{
\dontrun{
fin <- "ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/NOV/DNRES/DNRO2013.DBC"
fout <- tempfile(fileext = ".DBC")
download.file(fin, destfile = fout)
d <- read_datasus(fout)
}
}
