% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{sim_example}
\alias{sim_example}
\title{Example of DATASUS SIM Dataset}
\usage{
sim_example
}
\description{
Example of data structure of a datset from the DATASUS SIM database. Note that this is not real data. It is 100 records of variables independently sampled from plausible values. Combinations of variables for a given record may not make sense. The purpose of this data is to provide an example of the kind of structure the functions in this package deal with without providing real data.
}
