% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform.R
\name{transform_data}
\alias{transform_data}
\title{Transform DATASUS data according to a dictionary}
\usage{
transform_data(data, dict, keep = NULL, language = c("en", "br"),
  quiet = FALSE)
}
\arguments{
\item{data}{a data frame to transform}

\item{dict}{a dictionary to use for the transformation}

\item{keep}{an optional vector of column names to keep (others will be omitted)}

\item{language}{should the resulting transformed data be in English or Portuguese? Currently not implemented.}

\item{quiet}{logical - should messages be suppressed?}
}
\description{
Transform DATASUS data according to a dictionary
}
\examples{
d <- transform_data(sinasc_example, subset(cdcs_dict, db == "SINASC"))

\dontrun{
fin <- "ftp://ftp.datasus.gov.br/dissemin/publicos/SINASC/NOV/DNRES/DNRO2013.DBC"
fout <- tempfile(fileext = ".DBC")
download.file(fin, destfile = fout)
d <- read_datasus(fout)
d <- transform_data(d, subset(cdcs_dict, db == "SINASC"))
}
}
